devops/gcp/cloudfunction/helloworld/requirements.txt
psycopg2==2.7.5
PyYAML
